[iris] Move provider protocols and infra types from backends/ to platforms/
The ControllerProvider/WorkerInfraProvider protocols and the machine-
lifecycle types (handles, status enums, InfraError hierarchy, network
helpers) lived in cluster/backends/ while all their implementations
moved to cluster/platforms/ in #6727. Move backends/protocols.py and
backends/types.py to platforms/ and rewire every importer; backends/
now holds only the TaskBackend implementations (rpc/, k8s/). The
backends package re-exports are gone — importers use the defining
modules directly.

Also refresh the stale protocols.py docstring (still pointed at
providers/gcp/), document platforms/ in architecture.md — which still
drew machine lifecycle under backends/ — and correct its claim that
the decision kernels do no live I/O: the autoscaler actuates through
WorkerInfraProvider.

Part of #6914
