+++ lib/iris/src/iris/cluster/controller/autoscaler/worker_registry.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass, field
 
 from rigging.timing import Duration
 
-from iris.cluster.backends.types import (
+from iris.cluster.platforms.types import (
     CloudWorkerState,
     CommandResult,
     RemoteWorkerHandle,
