+++ lib/iris/src/iris/cluster/local_cluster.py
@@ -25,7 +25,6 @@ from rigging.credential_store import CredentialRecord, save_credentials
 from rigging.timing import Duration, Timestamp
 
 from iris.cluster.backends.rpc.backend import RpcTaskBackend, RpcWorkerStubFactory
-from iris.cluster.backends.types import find_free_port
 from iris.cluster.config import (
     GcpPlatformConfig,
     IrisClusterConfig,
@@ -51,6 +50,7 @@ from iris.cluster.controller.db import ControllerDB
 from iris.cluster.controller.log_stack import build_log_stack
 from iris.cluster.platforms.gcp.fake import InMemoryGcpService
 from iris.cluster.platforms.gcp.workers import GcpWorkerProvider
+from iris.cluster.platforms.types import find_free_port
 from iris.cluster.platforms.vm_lifecycle import ControllerStatus
 from iris.cluster.service_mode import ServiceMode
 from iris.cluster.types import DEFAULT_BACKEND_ID, AcceleratorType
