+++ lib/iris/src/iris/cluster/platforms/factory.py
@@ -5,7 +5,6 @@
 
 from dataclasses import dataclass
 
-from iris.cluster.backends.protocols import ControllerProvider, WorkerInfraProvider
 from iris.cluster.config import GcpPlatformConfig, IrisClusterConfig, PlatformConfig, SshConfig
 from iris.cluster.platforms.gcp.controller import GcpControllerProvider
 from iris.cluster.platforms.gcp.fake import InMemoryGcpService
@@ -13,6 +12,7 @@ from iris.cluster.platforms.gcp.workers import GcpWorkerProvider
 from iris.cluster.platforms.k8s.controller import K8sControllerProvider
 from iris.cluster.platforms.manual.controller import ManualControllerProvider
 from iris.cluster.platforms.manual.workers import ManualWorkerProvider
+from iris.cluster.platforms.protocols import ControllerProvider, WorkerInfraProvider
 from iris.cluster.service_mode import ServiceMode
