+++ lib/iris/src/iris/cluster/platforms/k8s/controller.py
@@ -21,7 +21,6 @@ import fsspec.config
 import s3fs
 from rigging.timing import Deadline
 
-from iris.cluster.backends.types import InfraError, Labels, local_queue_name
 from iris.cluster.config import ControllerVmConfig, CoreweavePlatformConfig, IrisClusterConfig, config_to_dict
 from iris.cluster.inject_env import TASK_ENV_SECRET_NAME, collect_inject_env, projects_task_env_secret
 from iris.cluster.platforms.k8s.constants import COREWEAVE_INTERRUPTABLE_TOLERATION, NVIDIA_GPU_TOLERATION
@@ -33,6 +32,7 @@ from iris.cluster.platforms.k8s.types import (
     K8sResource,
     parse_k8s_timestamp,
 )
+from iris.cluster.platforms.types import InfraError, Labels, local_queue_name
 
 logger = logging.getLogger(__name__)
