+++ lib/iris/src/iris/cluster/worker/task_attempt.py
@@ -23,10 +23,10 @@ from finelog.rpc import logging_pb2
 from rigging.timing import Duration, ExponentialBackoff, Timestamp
 
 from iris.chaos import chaos, chaos_raise
-from iris.cluster.backends.types import probe_outbound_ip
 from iris.cluster.bundle import BundleStore
 from iris.cluster.constraints import WellKnownAttribute
 from iris.cluster.log_keys import classify_log_level, task_log_key
+from iris.cluster.platforms.types import probe_outbound_ip
 from iris.cluster.runtime.docker import DockerContainerHandle
 from iris.cluster.runtime.env import build_common_iris_env
 from iris.cluster.runtime.types import (
