+++ lib/iris/tests/cluster/backends/gcp/test_gcp_service.py
@@ -17,13 +17,6 @@ from unittest.mock import MagicMock, patch
 
 import httpx
 import pytest
-from iris.cluster.backends.types import (
-    CloudSliceState,
-    InfraError,
-    InfraUnavailableError,
-    QuotaExhaustedError,
-    ResourceNotFoundError,
-)
 from iris.cluster.platforms.gcp.fake import InMemoryGcpService
 from iris.cluster.platforms.gcp.handles import GcpVmSliceHandle
 from iris.cluster.platforms.gcp.service import (
@@ -31,6 +24,13 @@ from iris.cluster.platforms.gcp.service import (
     TpuCreateRequest,
     VmCreateRequest,
     VmInfo,
+)
+from iris.cluster.platforms.types import (
+    CloudSliceState,
+    InfraError,
+    InfraUnavailableError,
+    QuotaExhaustedError,
+    ResourceNotFoundError,
 )
 from iris.cluster.service_mode import ServiceMode
 from iris.cluster.types import CapacityType
