+++ lib/iris/tests/cluster/controller/conftest.py
@@ -17,7 +17,6 @@ import pytest
 from finelog.client.log_client import Table
 from finelog.rpc.logging_connect import LogServiceClientSync
 from iris.cluster.backends.rpc.backend import WORKER_RECONCILE_TEARDOWN_REASON
-from iris.cluster.backends.types import CloudSliceState
 from iris.cluster.bundle import BundleStore
 from iris.cluster.config import (
     AutoscalerConfig,
@@ -91,6 +90,7 @@ from iris.cluster.controller.worker_health import (
 from iris.cluster.federation.manager import FederationManager
 from iris.cluster.platforms.gcp.fake import InMemoryGcpService
 from iris.cluster.platforms.gcp.workers import GcpWorkerProvider
+from iris.cluster.platforms.types import CloudSliceState
 from iris.cluster.service_mode import ServiceMode
 from iris.cluster.types import (
     DEFAULT_BACKEND_ID,
