+++ lib/iris/tests/cluster/controller/test_vm_lifecycle.py
@@ -30,12 +30,6 @@ from collections.abc import Callable
 from contextlib import AbstractContextManager, nullcontext
 
 import pytest
-from iris.cluster.backends.types import (
-    CloudWorkerState,
-    CommandResult,
-    SliceHandle,
-    WorkerStatus,
-)
 from iris.cluster.config import (
     ControllerVmConfig,
     GcpControllerConfig,
@@ -44,6 +38,12 @@ from iris.cluster.config import (
     PlatformConfig,
     SliceConfig,
     VmConfig,
+)
+from iris.cluster.platforms.types import (
+    CloudWorkerState,
+    CommandResult,
+    SliceHandle,
+    WorkerStatus,
 )
 from iris.cluster.platforms.vm_lifecycle import (
     _build_controller_vm_config,
