+++ lib/iris/tests/e2e/_docker_cluster.py
@@ -18,7 +18,6 @@ from finelog.rpc import logging_pb2
 from finelog.rpc.logging_connect import LogServiceClientSync
 from iris.client import IrisClient
 from iris.cluster.backends.rpc.backend import RpcTaskBackend, RpcWorkerStubFactory
-from iris.cluster.backends.types import find_free_port
 from iris.cluster.bundle import BundleStore
 from iris.cluster.config import (
     AutoscalerConfig,
@@ -35,6 +34,7 @@ from iris.cluster.config import (
 from iris.cluster.controller.controller import Controller, ControllerConfig
 from iris.cluster.controller.log_stack import build_log_stack
 from iris.cluster.local_cluster import LocalCluster
+from iris.cluster.platforms.types import find_free_port
 from iris.cluster.runtime.docker import DockerRuntime
 from iris.cluster.types import (
     DEFAULT_BACKEND_ID,
