+++ lib/iris/tests/e2e/gpu_gang_smoke.py
@@ -51,7 +51,6 @@ from pathlib import Path
 
 import click
 from iris.client import IrisClient
-from iris.cluster.backends.types import Labels, find_free_port
 from iris.cluster.config import CoreweavePlatformConfig, IrisClusterConfig, load_config
 from iris.cluster.platforms.k8s.controller import K8sControllerProvider, _build_controller_deployment
 from iris.cluster.platforms.k8s.coreweave_topology import (
@@ -63,6 +62,7 @@ from iris.cluster.platforms.k8s.coreweave_topology import (
     CW_LABEL_SUPERPOD,
 )
 from iris.cluster.platforms.k8s.service import CloudK8sService
+from iris.cluster.platforms.types import Labels, find_free_port
 from iris.cluster.types import AcceleratorType, CoschedulingConfig, Entrypoint, EnvironmentSpec, ResourceSpec, gpu_device
 from iris.rpc import job_pb2
